New structure of project

+++ src/pl/wybornie/dao/RoleDAO.java
@@ -4,6 +4,5 @@ import pl.wybornie.entity.Role;
 
 public interface RoleDAO {
 	
-	public Role getRole(int id);
-
+	public Role getRole(Long id);
 }
